moved scroll method to helpers module
moved scroll method to helpers module
