+++ tests/milo/html_ext.test.js
@@ -2,32 +2,10 @@ const { expect, test } = require('@playwright/test');
 const parse = require('../../libs/parse.js');
 const htmlExt = require('../../features/milo/html_ext.spec.js');
 const selectors = require('../../selectors/milo/html_ext.selectors.js');
+const scroll = require('../../libs/helpers.js');
 
 // Parse the feature file into something flat that can be tested separately
 const { name, features } = parse(htmlExt);
-
-/**
- * Slow/fast scroll JS evaluation method.
- * To be used in page.evaluate, i.e. page.evaluate(scroll, { direction: 'value', speed: 'value' });
- * @param direction string direction you want to scroll
- * @param speed string speed you would like to scroll. Options: slow, fast
-*/
-const scroll = async (args) => {
-  const { direction, speed } = args;
-  // eslint-disable-next-line no-promise-executor-return
-  const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-  const scrollHeight = () => document.body.scrollHeight;
-  const start = direction === 'down' ? 0 : scrollHeight();
-  const shouldStop = (position) => (direction === 'down' ? position > scrollHeight() : position < 0);
-  const increment = direction === 'down' ? 100 : -100;
-  const delayTime = speed === 'slow' ? 30 : 5;
-  console.error(start, shouldStop(start), increment);
-  for (let i = start; !shouldStop(i); i += increment) {
-    window.scrollTo(0, i);
-    // eslint-disable-next-line no-await-in-loop
-    await delay(delayTime);
-  }
-};
 
 test.describe(`${name}`, () => {
   features.forEach((props) => {
